Use "import/export" syntax for GGRC.extensions
