+++ src/ggrc-client/js/ggrc_base.js
@@ -3,11 +3,10 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-(function (GGRC, moment) {
-  GGRC.mustache_path = '/static/mustache';
+GGRC.mustache_path = '/static/mustache';
 
-  GGRC.extensions = GGRC.extensions || [];
-  if (!GGRC.widget_descriptors) {
-    GGRC.widget_descriptors = {};
-  }
-})(window.GGRC = window.GGRC || {}, moment);
+if (!GGRC.widget_descriptors) {
+  GGRC.widget_descriptors = {};
+}
+
+
